Fix image re-binding issues for contacts obtained via a LookupProvider
The images loaded via information from a LookupProvider, like those in
the main Converations view, don't rebind after the view undergoes
a detatchment and re-attchment phase. RecyclerView, unlike ListView,
doesn't call bind if it deems a ViewHolder in its scrap-heap is valid.
In such cases, the ViewHolder is attached to the view hierarchy without
undergoing a bind().

The ImageViews maintain a strong binding to model-data which enables them
to free the view-state on detachment. On attachment, the view-state is
refreshed from the model-state.

The bugs we were seeing were a result of not updating the model-state with
the image information obtained from the LookupProvider.

Change-Id: Id93ee6369435211f20d7bfe16d521c79669e4f4a
Issue-Id: CYNGNOS-2118

Conflicts:
	src/com/cyanogenmod/messaging/ui/WebImageView.java